Damageless Bonus
Adds a +1 bonus token if you manage to beat a round damageless.

Adds a new isDamageless variable in PlayerHealth that is false by default. This is accessed through the PortalScript, where it checks if you had a damageless round and rewards you.

Things that need improving:
getting the current round and setting this bonus to be after round 3. Currently, it is after round 1.
A damageless text after you go through the portal.